+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 0.23.0
+
 * feat(app-firewall-rules) Add support for application firewall rules
 
 ## 0.22.3